refactor: update api stuff
