+++ frontend/api/auth.ts
@@ -1,12 +1,11 @@
 import { API_URL } from "@/api/constants";
 import { redirect } from "next/navigation";
-import { NextRequest } from "next/server";
-import { cookies } from "next/headers";
 
-export async function POST(req: NextRequest) {
-  if (cookies().get("user")) {
+const login = async (req: any) => {
+  console.log(req);
+  /* if (cookies().get("user")) {
     cookies().delete("user");
-  }
+  } */
 
   const form = await req.formData();
   form.append("pub_date", new Date().toISOString());
@@ -28,7 +27,7 @@ export async function POST(req: NextRequest) {
       const segments = segment.split(";");
       const [keyValue] = segments;
       const [name, value] = keyValue.split("=").map((str) => str.trim());
-      if (name === "csrftoken" || name === "sessionid") {
+      /* if (name === "csrftoken" || name === "sessionid") {
         if (name === "csrftoken") {
           cookies().set("userToken", value);
         }
@@ -38,11 +37,11 @@ export async function POST(req: NextRequest) {
         }
 
         cookiesSet = true;
-      }
+      } */
     });
 
     if (cookiesSet) {
-      cookies().set("user", JSON.stringify(result));
+      /* cookies().set("user", JSON.stringify(result)); */
       redirect("/dashboard");
     } else {
       console.error("Failed to set cookies");
@@ -52,4 +51,19 @@ export async function POST(req: NextRequest) {
     console.error("No cookies received from the server");
     redirect("/login");
   }
-}
+};
+
+const logout = async (req: any) => {
+  const res = await fetch(`${API_URL}/logout`, { ...req, method: "POST" });
+
+  /* cookies().delete("userToken");
+  cookies().delete("userSession");
+  cookies().delete("user"); */
+
+  redirect("/");
+};
+
+export const authApi = {
+  userLogin: login,
+  userLogout: logout,
+};
