+++ frontend/api/index.ts
@@ -1,3 +1,5 @@
 export * from "./products";
 export * from "./pets";
 export * from "./user";
+export * from "./auth";
+export * from "./store";
