+++ frontend/components/CheckoutButton/CheckoutButton.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React, { useState, useEffect } from "react";
 import styles from "./CheckoutButton.module.css";
 
@@ -8,7 +6,7 @@ export const CheckoutButton = (props: any) => {
   const [url, setUrl] = useState<null | string>(null);
   const [loading, setLoading] = useState<boolean>(true);
 
-  useEffect(() => {
+  /* useEffect(() => {
     const generateLink = async () => {
       setLoading(true);
 
@@ -26,7 +24,7 @@ export const CheckoutButton = (props: any) => {
     };
 
     generateLink();
-  }, [product]);
+  }, [product]); */
 
   return (
     <div>
