+++ frontend/components/Header/Header.tsx
@@ -2,7 +2,8 @@ import { cookies } from "next/headers";
 import styles from "./Header.module.css";
 import { Navigation } from "../Navigation";
 
-export const Header = () => {
+export const Header = (props: any) => {
+  const { type } = props;
   return (
     <header>
       <a className={styles.skipMain} href="#main">
