+++ frontend/pages/_document.tsx
@@ -1,5 +1,4 @@
 import { Html, Head, Main, NextScript } from "next/document";
-import "bootstrap/dist/css/bootstrap.min.css";
 
 export default function Document() {
   return (
