+++ frontend/pages/dashboard/pets/[id]/index.tsx
@@ -1,4 +1,4 @@
-import styles from "./page.module.css";
+import styles from "./index.module.css";
 import { petsApi } from "@/api";
 
 const Pet = async (props: any) => {
