+++ frontend/pages/dashboard/products/[id]/index.tsx
@@ -1,4 +1,4 @@
-import styles from "./page.module.css";
+import styles from "./index.module.css";
 import { productsApi } from "@/api";
 
 const Product = async (props: any) => {
